--- resourses/fxml/addStudent.fxml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-
-
-<AnchorPane maxHeight="500.0" maxWidth="1000.0" minHeight="500.0" minWidth="1000.0" prefHeight="500.0" prefWidth="1000.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controllers.AddStudent">
-   <children>
-      <TextField fx:id="surname" layoutX="80.0" layoutY="15.0" promptText="фамилия" />
-      <TextField fx:id="middleName" layoutX="80.0" layoutY="75.0" promptText="отчество" />
-      <TextField fx:id="name" layoutX="80.0" layoutY="45.0" promptText="имя" />
-      <ChoiceBox fx:id="choiceGroup" layoutX="80.0" layoutY="105.0" prefWidth="171.0" />
-      <Button fx:id="add" layoutX="80.0" layoutY="135.0" mnemonicParsing="false" onAction="#clickAdd" prefHeight="26.0" prefWidth="171.0" text="добавить" />
-      <Button fx:id="exit" layoutX="80.0" layoutY="460.0" mnemonicParsing="false" onAction="#clickExit" prefHeight="26.0" prefWidth="171.0" text="применить и выйти" />
-      <TableView fx:id="table" layoutX="336.0" layoutY="14.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="469.0" prefWidth="650.0" AnchorPane.bottomAnchor="15.0" AnchorPane.leftAnchor="335.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="15.0">
-         <columns>
-            <TableColumn fx:id="choice" maxWidth="100.0" minWidth="100.0" prefWidth="100.0" text="выбор" />
-            <TableColumn fx:id="surnameTable" prefWidth="120.0" text="фамллия" />
-            <TableColumn fx:id="nameTable" prefWidth="150.0" text="имя" />
-            <TableColumn fx:id="middleNameTable" prefWidth="112.0" text="отчество" />
-            <TableColumn fx:id="group" maxWidth="81.0" minWidth="80.0" prefWidth="81.0" text="группа" />
-         </columns>
-         <columnResizePolicy>
-            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-         </columnResizePolicy>
-      </TableView>
-   </children>
-</AnchorPane>
