--- resourses/fxml/editListStudent.fxml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane prefHeight="700.0" prefWidth="1000.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controllers.EditListStudent">
-   <children>
-      <TableView fx:id="table" layoutX="336.0" layoutY="10.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="680.0" prefWidth="650.0" AnchorPane.bottomAnchor="10.0" AnchorPane.leftAnchor="340.0" AnchorPane.rightAnchor="10.0" AnchorPane.topAnchor="10.0">
-        <columns>
-          <TableColumn fx:id="choice" maxWidth="100.0" minWidth="100.0" prefWidth="100.0" text="выбор" />
-          <TableColumn fx:id="surname" prefWidth="120.0" text="фамллия" />
-            <TableColumn fx:id="name" prefWidth="150.0" text="имя" />
-            <TableColumn fx:id="middleName" prefWidth="112.0" text="отчество" />
-            <TableColumn fx:id="group" maxWidth="81.0" minWidth="80.0" prefWidth="81.0" text="группа" />
-        </columns>
-         <columnResizePolicy>
-            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-         </columnResizePolicy>
-      </TableView>
-      <Button fx:id="addStudent" layoutX="60.0" layoutY="50.0" mnemonicParsing="false" onAction="#clickAddStudent" prefHeight="26.0" prefWidth="200.0" text="добавить студента" AnchorPane.leftAnchor="60.0" AnchorPane.topAnchor="50.0" />
-      <Button fx:id="exit" layoutX="60.0" layoutY="650.0" mnemonicParsing="false" onAction="#exit" prefHeight="26.0" prefWidth="200.0" text="выход" AnchorPane.bottomAnchor="25.0" AnchorPane.leftAnchor="60.0" />
-      <ChoiceBox fx:id="choiceGroup" layoutX="60.0" layoutY="110.0" prefWidth="200.0" AnchorPane.leftAnchor="60.0" AnchorPane.topAnchor="110.0" />
-      <Button fx:id="addGroup" layoutX="60.0" layoutY="80.0" mnemonicParsing="false" onAction="#clickAddGroup" prefHeight="26.0" prefWidth="200.0" text="добавить группу" AnchorPane.leftAnchor="60.0" AnchorPane.topAnchor="80.0" />
-   </children>
-</AnchorPane>
