--- resourses/fxml/mainWindow.fxml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane prefHeight="760.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controllers.MainWindow">
-    <children>
-        <Text fx:id="nameGroup" layoutX="641.0" layoutY="46.0" strokeWidth="0.0" text="группа" textAlignment="CENTER" textOrigin="CENTER" wrappingWidth="350.0">
-            <font>
-                <Font size="40.0" />
-            </font>
-        </Text>
-
-        <Button fx:id="listGroup" layoutX="14.0" layoutY="96.0" mnemonicParsing="false" onAction="#editListStudent" prefHeight="30.0" prefWidth="275.0" text="редактировать группы" />
-      <ComboBox fx:id="choiceGroup" layoutX="77.0" layoutY="171.0" prefWidth="150.0" />
-        <TableView fx:id="table" cache="true" editable="true" fixedCellSize="0.0" layoutX="300.0" layoutY="96.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="660.0" prefWidth="975.0" AnchorPane.bottomAnchor="4.0" AnchorPane.leftAnchor="300.0" AnchorPane.rightAnchor="4.0" AnchorPane.topAnchor="96.0">
-            <columns>
-                <TableColumn fx:id="idLine" maxWidth="30.0" minWidth="30.0" prefWidth="30.0" text="№" />
-                <TableColumn fx:id="surname" maxWidth="500.0" minWidth="100.0" prefWidth="120.0" text="Фамилия" />
-                <TableColumn fx:id="name" maxWidth="500.0" minWidth="100.0" prefWidth="120.0" text="имя" />
-                <TableColumn fx:id="middleName" maxWidth="500.0" minWidth="100.0" prefWidth="120.0" text="отчество" />
-                <TableColumn fx:id="grades" maxWidth="600.0" minWidth="300.0" prefWidth="400.0" text="оценки" />
-                <TableColumn fx:id="averageScore" maxWidth="85.0" minWidth="85.0" prefWidth="85.0" text="средняя" />
-            </columns>
-         <columnResizePolicy>
-            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-         </columnResizePolicy>
-        </TableView>
-    </children>
-
-</AnchorPane>
