--- src/controllers/ConfigsControllers.java
@@ -1,10 +1,0 @@
-package controllers;
-
-
-import javafx.event.ActionEvent;
-import java.io.IOException;
-
-public abstract class ConfigsControllers {
-
-    public abstract void createWindow(ActionEvent actionEvent) throws IOException;
-}
