--- src/dao/gradeDao/GradeDao.java
@@ -1,11 +1,0 @@
-package dao.gradeDao;
-
-import entity.model.Student;
-
-import java.util.List;
-
-public abstract class GradeDao {
-    public abstract List findGrade(int idStudent);
-    public abstract void editGrade(Student student, List<Integer> gradeList);
-    public abstract void deleteFullGrades(int idStudent);
-}
