--- src/dao/studentDao/StudentDao.java
@@ -1,13 +1,0 @@
-package dao.studentDao;
-
-import entity.model.Grade;
-import entity.model.Student;
-import entity.model.StudentGroup;
-import entity.view.StudentView;
-
-import java.util.List;
-
-public abstract class StudentDao {
-    public abstract List<Student> findStudents();
-    public abstract void deleteStudent(int idStudent);
-}
