--- src/dao/studentGroupDao/StudentGroupDao.java
@@ -1,10 +1,0 @@
-package dao.studentGroupDao;
-
-import entity.model.StudentGroup;
-
-import java.util.List;
-
-public abstract class StudentGroupDao {
-    public abstract List findGroup();
-    public abstract StudentGroup findGroupFromId(int id);
-}
