+++ src/main/java/utility/dataBase/ConfigDB.java
@@ -1,0 +1,8 @@
+package utility.dataBase;
+
+public class ConfigDB {
+    public static final String NAME_DB = "student_db";
+    public static final String URL = "jdbc:mysql://localhost/" + NAME_DB;
+    public static final String USER = "root";
+    public static final String PASSWORD = "root";
+}
