+++ src/main/resources/fxml/addStudent.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane maxHeight="155.0" maxWidth="200.0" minHeight="155.0" minWidth="200.0" prefHeight="155.0" prefWidth="200.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controllers.AddStudentController">
+   <children>
+      <TextField fx:id="txtSurname" layoutX="14.0" layoutY="9.0" promptText="фамилия" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="5.0" />
+      <TextField fx:id="txtMiddleName" layoutX="14.0" layoutY="64.0" promptText="отчество" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="65.0" />
+      <TextField fx:id="txtName" layoutX="14.0" layoutY="39.0" promptText="имя" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="35.0" />
+      <ChoiceBox fx:id="choiceGroup" layoutX="3.0" layoutY="100.0" prefWidth="171.0" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="95.0" />
+      <Button fx:id="add" layoutX="5.0" layoutY="120.0" mnemonicParsing="false" onAction="#clickAdd" prefHeight="26.0" prefWidth="171.0" text="добавить" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="125.0" />
+   </children>
+</AnchorPane>
