+++ src/main/resources/fxml/mainWindow.fxml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.*?>
+
+<AnchorPane prefHeight="760.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controllers.MainWindowController">
+    <children>
+        <Text fx:id="nameGroup" layoutX="641.0" layoutY="46.0" strokeWidth="0.0" text="все группы" textAlignment="CENTER" textOrigin="CENTER" wrappingWidth="350.0">
+            <font>
+                <Font size="40.0" />
+            </font>
+        </Text>
+
+        <Button fx:id="listGroup" layoutX="15.0" layoutY="14.0" mnemonicParsing="false" onAction="#buttonEditListStudent" prefHeight="30.0" prefWidth="275.0" text="редактировать списки" />
+      <ComboBox fx:id="choiceGroup" layoutX="78.0" layoutY="46.0" prefWidth="150.0" />
+        <TableView fx:id="table" cache="true" editable="true" fixedCellSize="0.0" layoutX="-4.0" layoutY="96.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="660.0" prefWidth="1280.0" AnchorPane.bottomAnchor="5.0" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="100.0">
+            <columns>
+                <TableColumn fx:id="idLine" maxWidth="30.0" minWidth="30.0" prefWidth="30.0" text="№" />
+                <TableColumn fx:id="surname" maxWidth="200.0" minWidth="100.0" prefWidth="120.0" text="Фамилия" />
+                <TableColumn fx:id="name" maxWidth="200.0" minWidth="120.0" prefWidth="120.0" text="имя" />
+                <TableColumn fx:id="middleName" maxWidth="200.0" minWidth="100.0" prefWidth="120.0" text="отчество" />
+                <TableColumn fx:id="grades" maxWidth="600.0" minWidth="300.0" prefWidth="400.0" text="оценки" />
+                <TableColumn fx:id="averageScore" maxWidth="85.0" minWidth="85.0" prefWidth="85.0" text="средняя" />
+            </columns>
+         <columnResizePolicy>
+            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+         </columnResizePolicy>
+        </TableView>
+    </children>
+
+</AnchorPane>
