--- src/utility/dataBase/ConfigDB.java
@@ -1,8 +1,0 @@
-package utility.dataBase;
-
-public class ConfigDB {
-    public static final String NAME_DB = "students_db";
-    public static final String URL = "jdbc:mysql://localhost/" + NAME_DB;
-    public static final String USER = "root";
-    public static final String PASSWORD = "root";
-}
